requirements.txt
discord.py
langchain
faiss-cpu
transformers
sentence-transformers
